Added semi-finished Graph ADT
The Graph ADT has been added to the project. Note the interface and
actual class methods are not implemented correctly, this is just a
starting point to build off of. Consider reviewing the slides for an
interface that shows how to actually construct the thing. Possibly move
the makeGrid function to the main method since it's not portable.